Add integration tests
Not sure how to do them, but at least we already got an issue with the
Finish button label when disabled.

The expectations for the results are using a nested HTML path, which is
not great to maintain. I leave this way to try a refactor in another
commit, maybe with an attribute `id`.